+++ llrt_modules/src/lib.rs
@@ -3,6 +3,7 @@
 #![allow(clippy::new_without_default)]
 pub use self::module_info::ModuleInfo;
 pub use self::modules::*;
+use std::sync::atomic::AtomicUsize;
 
 mod module_info;
 mod modules;
@@ -11,3 +12,7 @@ pub mod stream;
 #[cfg(test)]
 mod test;
 mod utils;
+
+pub const VERSION: &str = env!("CARGO_PKG_VERSION");
+
+pub static TIME_ORIGIN: AtomicUsize = AtomicUsize::new(0);
